+++ apps/console/src/ui/pages/SkillsPage.tsx
@@ -2,5 +2,5 @@ import React from 'react';
 import { DiscoverPage } from './DiscoverPage';
 
 export function SkillsPage(): JSX.Element {
-  return <DiscoverPage mode="skills" />;
+  return <DiscoverPage />;
 }